move avg_degree() to basic/avg_degree.py
