+++ easygraph/functions/basic/__init__.py
@@ -1,3 +1,4 @@
+from .avg_degree import *
 from .cluster import *
 from .laplacian import *
 from .pagerank import *
